+++ lint-staged.config.ts
@@ -1,6 +1,8 @@
 const lintStagedConfig = {
   // function here prevents file names appended to the script
-  "*": () => "pnpm check",
+  "*": () => {
+    return "pnpm check";
+  },
 };
 
 export default lintStagedConfig;