+++ src/configs/oxlint/index.ts
@@ -1,3 +1,0 @@
-export { default as base } from "./base.ts";
-export { default as lib } from "./lib.ts";
-export { default as next } from "./next.ts";