+++ src/lib/index.ts
@@ -1,4 +1,0 @@
-export { throwError, log, logDivider, logNewLine, warn, warnNewLine } from "./logger.ts";
-export { deepCopy, deepMergeObj, freshRegex } from "./operations.ts";
-export { isArr, isFn, isNum, isObj, isStr } from "./types.ts";
-export { debounce, sleep } from "./utils.ts";