+++ src/plugins/vite/index.ts
@@ -1,3 +1,0 @@
-// oxlint-disable import/prefer-default-export
-
-export { default as copyFolders } from "./copy-folders.ts";